Add nodes for multiplexing and demultiplexing
The current implementation works in a deterministic way, when
multiplexing, first a packet from the first input is taken then from the
second input and so on. The demultiplexing works in a similar way.

For the future it may be a good idea to provide multiple strategies that
the user can choose from, e.g. a greedy strategy for the multiplexer.