Introduce the TunTapNode
This node is a sink type node that allows to exfiltrate traffic to a TUN
or TAP interface. This functionionality also required some restructuring
in the packet processing logic.

Sadly, at least currently, using this node requires the program to be
run as root. In addition it may be necessary to temporarily allow the
root to connect to the X11 server. The scripts introduced with this
commit can be used to add and later on remove these priviliges for root.